display the release command output if it is available
tell the customer to retry with `releases:output` if we got a 404

It can take a few seconds before the stream is actually available. This
message is there until we fix the issue with retries.

set the retry variable by equally directly from yield

don't overcomplicate things with a non-reused lib file